Company: small fixes

+++ OOP_Polimorphism06/src/company/employee/Employee.java
@@ -4,5 +4,5 @@ public abstract class Employee {
     private String firstName;
     private String lastName;
 
-    public abstract double salaryCalc();
+    public abstract float salaryCalc();
 }
